requirements.txt
# MCGL Project Dependencies
# Install with: pip install -r requirements.txt

# Core ML
torch>=2.0.0
torchvision
torchaudio
torch-geometric>=2.4.0
pykeen>=1.10.0

# Biomedical
TxGNN
transformers>=4.30.0
rdkit  # molecular fingerprints

# Data & Science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
networkx>=3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# RAG (for KGQA baseline)
langchain>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Experiment tracking
wandb>=0.15.0

# Configuration
pyyaml>=6.0
hydra-core>=1.3.0

# Utilities
tqdm>=4.65.0
rich>=13.0.0
